fix: MySQL→达梦整数与文本类型映射改用原生类型
- tinyint/smallint/int/bigint 改为达梦对应的 TINYINT/SMALLINT/INT/BIGINT，
  不再用 NUMBER(n) 兜底；mediumint 达梦无对应类型，用 INT 承接
- text/mediumtext/longtext 改为 TEXT，不再用 CLOB
- 同步更新黄金测试 dameng_test.go 及 mysql_dameng_test.go 的期望值

+++ datamigrate/dialect/dameng_test.go
@@ -24,9 +24,9 @@ func TestDaMengCreateTable_Golden(t *testing.T) {
 	}
 	want := `DROP TABLE "APP"."users" CASCADE CONSTRAINTS;
 CREATE TABLE "APP"."users" (
-  "id" NUMBER(19) IDENTITY(1, 1) NOT NULL,
+  "id" BIGINT IDENTITY(1, 1) NOT NULL,
   "name" VARCHAR2(100) NOT NULL,
-  "age" NUMBER(10) DEFAULT '0',
+  "age" INT DEFAULT '0',
   "created" TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 )`
 	if got := JoinSQL(stmts); got != want {
